Move compound command parser functions
